15% speed

+++ user/include/control.h
@@ -21,7 +21,7 @@ extern PID_Calibration PID_out_left;
 extern PID_Calibration PID_out_right;
 
 void Update_Motors(PID_State * left_state,PID_State * right_state,PID_State * gyro_state);
-void Set_Motors(unsigned int left, unsigned int right);
+void Set_Motors(long left, long right);
 void Speed_Ctrl_in(unsigned int accy_target);
 void Speed_Ctrl_mid(unsigned int adc_target);
 void Speed_Ctrl_out(unsigned int Left_Speed,unsigned int Right_Speed);
